refactor structure and exit cleanly
