+++ package.json
@@ -2,9 +2,12 @@
   "name": "gmmk-driver",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "./dist/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "npx tsc",
+    "start": "node ./dist/index.js",
+    "dev": "ts-node-dev --transpile-only ./src/index.ts"
   },
   "keywords": [],
   "author": "",
@@ -13,5 +16,8 @@
     "ts-node": "^10.5.0",
     "typescript": "^4.5.5",
     "usb": "^2.1.2"
+  },
+  "devDependencies": {
+    "ts-node-dev": "^1.1.8"
   }
 }
