+++ src/utils.ts
@@ -1,0 +1,1 @@
+export const delay = (ms: number) => new Promise<void>(resolve => setTimeout(resolve, ms));
